+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - [FP-2152 - Create new editor will close editor window](https://movai.atlassian.net/browse/FP-2152)
 - [FP-2128 - Add container configurations in Node editor](https://movai.atlassian.net/browse/FP-2147)
 - [FP-2186 - Main Menu tools are not showing up](https://movai.atlassian.net/browse/FP-2186)
+- [FP-2221 - Disable the Language server](https://movai.atlassian.net/browse/FP-2221)